Task #458
Added the reference, description and remarks fields to the jobs table, which refer to the product that will be used to bill the customer for the job.
Changed the field name "numhoras" by quantity, which will reflect the quantity to be invoiced to the client for the work.
Added or revised files for translations.
Added to the controller an action button to autocalculate the hours worked based on the data of the work performed.